Use react-motion
