+++ package.json
@@ -23,24 +23,36 @@
     "@kadira/storybook": "^2.35.3",
     "@types/bluebird": "^3.0.37",
     "@types/chai": "^3.4.33",
+    "@types/classnames": "^0.0.32",
     "@types/enzyme": "^2.7.4",
     "@types/history": "^4.5.0",
+    "@types/jest": "^18.1.1",
     "@types/jsdom": "^2.0.29",
+    "@types/node": "^7.0.5",
+    "@types/react": "^15.0.8",
+    "@types/react-addons-css-transition-group": "^15.0.1",
+    "@types/react-dnd": "^2.0.32",
     "@types/react-dom": "^0.14.23",
+    "@types/react-redux": "^4.4.36",
     "@types/react-router": "^3.0.2",
     "@types/react-router-redux": "^4.0.40",
+    "@types/react-tabs": "^0.5.22",
+    "@types/redux": "^3.6.31",
+    "@types/redux-actions": "^1.2.2",
     "@types/redux-thunk": "^2.1.32",
     "@types/sinon": "^1.16.35",
     "autoprefixer": "^6.7.2",
     "babel-polyfill": "^6.22.0",
     "bluebird": "^3.4.7",
     "chai": "^3.5.0",
+    "classnames": "^2.2.5",
     "cpx": "^1.3.1",
     "css-loader": "^0.26.1",
     "electron": "^1.4.15",
     "enzyme": "^2.7.1",
     "filesaver.js": "^0.2.0",
     "html-webpack-plugin": "^2.28.0",
+    "immutable": "^3.8.1",
     "imports-loader": "^0.7.0",
     "jest": "^18.1.0",
     "json-loader": "^0.5.4",
@@ -49,48 +61,36 @@
     "node-sass": "^4.5.0",
     "npm-run-all": "^4.0.1",
     "postcss-loader": "^1.2.2",
+    "react": "^15.4.2",
+    "react-addons-css-transition-group": "^15.4.2",
     "react-addons-test-utils": "^15.4.2",
+    "react-dnd": "^2.2.3",
     "react-dnd-html5-backend": "^2.2.3",
     "react-dom": "^15.4.2",
     "react-hot-loader": "^1.3.0",
-    "react-router": "^3.0.2",
-    "react-router-redux": "^4.0.2",
-    "redux-thunk": "^2.2.0",
-    "sass-loader": "^5.0.1",
-    "sinon": "2.0.0-pre",
-    "skeleton-css": "^2.0.4",
-    "style-loader": "^0.13.1",
-    "ts-loader": "^2.0.0",
-    "ts-node": "^2.0.0",
-    "wallaby-webpack": "^0.0.33",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.3.0",
-    "@types/classnames": "^0.0.32",
-    "@types/node": "^7.0.5",
-    "@types/jest": "^18.1.1",
-    "@types/react": "^15.0.8",
-    "@types/react-addons-css-transition-group": "^15.0.1",
-    "@types/react-dnd": "^2.0.32",
-    "@types/react-redux": "^4.4.36",
-    "@types/react-tabs": "^0.5.22",
-    "@types/redux": "^3.6.31",
-    "@types/redux-actions": "^1.2.2",
-    "classnames": "^2.2.5",
-    "react": "^15.4.2",
-    "react-addons-css-transition-group": "^15.4.2",
-    "react-dnd": "^2.2.3",
     "react-icons": "^2.2.3",
     "react-keydown": "^1.6.2",
     "react-redux": "^5.0.2",
+    "react-router": "^3.0.2",
+    "react-router-redux": "^4.0.2",
     "react-simple-dropdown": "^1.1.4",
     "react-tabs": "^0.8.2",
     "redux": "^3.3.1",
     "redux-actions": "^1.2.1",
+    "redux-thunk": "^2.2.0",
     "rimraf": "^2.5.4",
-    "immutable": "^3.8.1",
+    "sass-loader": "^5.0.1",
+    "sinon": "2.0.0-pre",
+    "skeleton-css": "^2.0.4",
+    "style-loader": "^0.13.1",
     "treeify": "^1.0.1",
+    "ts-loader": "^2.0.0",
+    "ts-node": "^2.0.0",
     "tslint": "^4.4.2",
-    "typescript": "^2.1.6"
+    "typescript": "^2.1.6",
+    "wallaby-webpack": "^0.0.33",
+    "webpack": "^2.2.1",
+    "webpack-dev-server": "^2.3.0"
   },
   "jest": {
     "moduleFileExtensions": [
