+++ packages/dag-history-component/package.json
@@ -32,6 +32,7 @@
     "@types/react": "^15.0.8",
     "@types/react-addons-css-transition-group": "^15.0.1",
     "@types/react-dnd": "^2.0.32",
+    "@types/react-motion": "^0.0.21",
     "@types/react-redux": "^4.4.36",
     "@types/react-tabs": "^0.5.22",
     "@types/redux": "^3.6.31",
@@ -43,6 +44,7 @@
     "react-dnd": "^2.2.3",
     "react-icons": "^2.2.3",
     "react-keydown": "^1.6.2",
+    "react-motion": "^0.4.7",
     "react-redux": "^5.0.2",
     "react-simple-dropdown": "^1.1.4",
     "react-tabs": "^0.8.2",
