+++ packages/dag-history-component/src/components/BranchList/index.tsx
@@ -8,12 +8,14 @@ export interface IBranchListProps {
   activeBranch?: number;
   branches: IBranchProps[];
   onBranchClick?: (branchId: number) => void;
+  style?: any;
 }
 
 const BranchList: React.StatelessComponent<IBranchListProps> = ({
   activeBranch,
   branches,
   onBranchClick,
+  style,
 }) => {
   const branchViews = branches.map(s => (
     <Branch
@@ -24,7 +26,7 @@ const BranchList: React.StatelessComponent<IBranchListProps> = ({
     />
   ));
   return (
-    <div className="history-branch-list">
+    <div className="history-branch-list" style={{...style}}>
       <div className="history-branch-list-inner">
         {branchViews}
       </div>
@@ -36,6 +38,7 @@ BranchList.propTypes = {
   activeBranch: React.PropTypes.number,
   branches: React.PropTypes.arrayOf(React.PropTypes.shape(Branch.propTypes)),
   onBranchClick: React.PropTypes.func,
+  style: React.PropTypes.object,
 };
 
 export default BranchList;
