+++ packages/dag-history-component/src/components/Continuation/Continuation.scss
@@ -11,7 +11,6 @@ $grey: #797d7f;
   font-weight: 600;
   height: 14px;
   line-height: 14px;
-  min-height: 14px;
   min-width: 14px;
   padding: 2px;
   text-align: center;
