+++ packages/dag-history-component/src/components/State/State.scss
@@ -11,7 +11,6 @@ $charcoal: #5d6d7e;
   cursor: pointer;
   display: flex;
   height: 30px;
-  min-height: 30px;
   min-width: 250px;
   padding: 8px;
   position: relative;
