+++ packages/dag-history-component/src/components/State/index.tsx
@@ -1,47 +1,154 @@
 import * as React from 'react';
-import State from './State';
-import { IExpandableStateProps } from './interfaces';
-import { IContinuationProps } from '../Continuation';
-import * as ReactCSSTransitionGroup from 'react-addons-css-transition-group';
+import * as classnames from 'classnames';
+import Continuation from '../Continuation';
+import colors from '../../palette';
+import './State.scss';
+import { IStateProps } from './interfaces';
+import { Motion, spring } from 'react-motion';
 
-const ExpandableState: React.StatelessComponent<IExpandableStateProps> = (props) => {
-  const {
-    id,
-    pinned,
-    active,
-    bookmarked,
-    branchType,
-    historyGraph,
-    getSourceFromState,
-    onClick,
-    onContinuationClick,
-    onBookmarkClick,
-    renderBookmarks,
-    successor,
-  } = props;
+const Bookmark = require('react-icons/lib/io/bookmark');
+const { PropTypes } = React;
 
-  const children = historyGraph.childrenOf(id);
-  const getSourceAndLabel = (stateId: number) => {
-    const state = historyGraph.getState(id);
-    const source = getSourceFromState(state);
-    const label = historyGraph.stateName(id);
-    return { source, label };
-  };
-
-  const stateProps = {
-    id,
-    pinned,
-    active,
-    bookmarked,
-    successor,
-    branchType,
-    onClick,
-    onContinuationClick,
-    onBookmarkClick,
-    numChildren: children.length,
-    renderBookmarks,
-    ...getSourceAndLabel(id),
-  };
-  return (<State {...stateProps} />);
+const coloring = {
+  current: {
+    active: colors.CURRENT_ACTIVE,
+    nonactive: colors.CURRENT,
+  },
+  legacy: {
+    active: colors.LEGACY_ACTIVE,
+    nonactive: colors.ANCESTOR,
+  },
+  unrelated: {
+    active: colors.UNRELATED,
+    nonactive: colors.UNRELATED_UNIQUE,
+  }
 };
-export default ExpandableState;
+
+function getBackgroundColor(branchType, active) {
+  let result = null;
+  result = coloring[branchType][active ? 'active' : 'nonactive'];
+  return result;
+}
+
+function continuationColor(isActive, isPinned) {
+  let result = colors.CONT_BLANK;
+  if (isPinned) {
+    result = colors.CONT_PINNED;
+  } else if (isActive) {
+    result = colors.CONT_ACTIVE;
+  }
+  return result;
+}
+
+export interface IStateState {}
+
+export default class State extends React.PureComponent<IStateProps, IStateState> {
+  static propTypes = {
+    id: PropTypes.number.isRequired,
+    source: PropTypes.string,
+    label: PropTypes.string.isRequired,
+    active: PropTypes.bool,
+    pinned: PropTypes.bool,
+    bookmarked: PropTypes.bool,
+    renderBookmarks: PropTypes.bool,
+    numChildren: PropTypes.number,
+    branchType: PropTypes.oneOf(['current', 'legacy', 'unrelated']),
+    onBookmarkClick: PropTypes.func,
+    onClick: PropTypes.func,
+    onContinuationClick: PropTypes.func,
+    showContinuation: PropTypes.bool,
+    style: PropTypes.any,
+  };
+
+  static defaultProps = {
+    id: undefined,
+    source: undefined,
+    showContinuation: true,
+    label: '',
+    branchType: 'current',
+    numChildren: 0,
+  };
+
+  public render() {
+    const {
+      id,
+      source,
+      label,
+      branchType,
+      active,
+      renderBookmarks,
+      bookmarked,
+      numChildren,
+      onClick,
+      onContinuationClick,
+      onBookmarkClick,
+      successor,
+      pinned,
+    } = this.props;
+
+    const backgroundColor = getBackgroundColor(branchType, active);
+
+    const handleClick = (e) => {
+      if (onClick) {
+        onClick(id);
+      }
+    };
+
+    const handleContinuationClick = (e) => {
+      if (onContinuationClick) {
+        onContinuationClick(id);
+      }
+    };
+
+    const handleBookmarkClick = (e) => {
+      if (onBookmarkClick) {
+        onBookmarkClick(id);
+      }
+    };
+
+    const continuation = !successor ? (
+      <Continuation
+        count={numChildren}
+        color={continuationColor(active, pinned)}
+        onClick={(e) => handleContinuationClick(e)}
+      />
+    ) : null;
+
+    const bookmark = renderBookmarks ? (
+      <Bookmark
+        size={25}
+        color={bookmarked ? 'gold' : 'white'}
+        onClick={e => handleBookmarkClick(e)}
+      />
+    ) : null;
+
+    const marginLeftValue = successor ? 30 : 0;
+
+    return (
+      <Motion style={{marginLeft: spring(marginLeftValue)}}>
+        {interpolatingStyle => (
+          <div
+            className={classnames('history-state', { successor })}
+            style={{
+              ...this.props.style,
+              backgroundColor,
+              marginLeft: interpolatingStyle.marginLeft
+            }}
+            onClick={e => handleClick(e)}
+          >
+            {continuation}
+            <div className="state-detail">
+              <div className={classnames('state-source', { active })}>
+                {source || ''}
+              </div>
+              <div className={classnames('state-name', { active })}>
+                {label || ''}
+              </div>
+            </div>
+            {bookmark}
+          </div>
+        )}
+      </Motion>
+    );
+  }
+}
